fix: 修复资源部录入界面展示
