+++ src/components/ElectronicTable/data.type.ts
